# docs/solomon_gpt_action_openapi.yaml
openapi: 3.0.3
info:
  title: SolomonGPT Action API
  description: Public API for connecting SolomonGPT Custom GPT with the Solomon Action Proxy and Command Center on SS1.
  version: 1.0.0
servers:
  - url: https://REPLACE_WITH_YOUR_PUBLIC_TUNNEL_URL/api
    description: Publicly reachable Solomon Action proxy/tunnel url
paths:
  /health:
    get:
      summary: Get proxy health status
      operationId: getHealth
      responses:
        '200':
          description: Health details of the proxy and backend
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  service:
                    type: string
                  status:
                    type: string

  /command-center/status:
    get:
      summary: Get command center system status
      operationId: getCommandCenterStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Overall status of the command center
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  status:
                    type: string
                  degraded:
                    type: boolean

  /command-center/bridge-status:
    get:
      summary: Get status of the backend SS1 bridge connections
      operationId: getBridgeStatus
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Connection status of the command center backend bridge
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  bridge_connected:
                    type: boolean
                  latency_ms:
                    type: integer

  /command-center/solomon-chat:
    post:
      summary: Send a message to Solomon and retrieve responses with Mnemosyne context injection
      operationId: sendSolomonChat
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - message
                - conversation_id
                - request_id
              properties:
                message:
                  type: string
                  description: The user prompt or message.
                conversation_id:
                  type: string
                  description: Stable conversation identifier.
                request_id:
                  type: string
                  description: Unique request identifier.
                security_classification:
                  type: string
                  enum: [PUBLIC, INTERNAL, RESTRICTED]
                  default: INTERNAL
                  description: Security clearance of the request.
                metadata:
                  type: object
                  properties:
                    source:
                      type: string
                      default: solomon-gpt
                  additionalProperties: true
      responses:
        '200':
          description: Coherent Solomon response along with memory telemetry
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reply:
                    type: string
                    description: Response generated by Solomon planner/runtime.
                  request_id:
                    type: string
                  conversation_id:
                    type: string
                  memory:
                    type: object
                    properties:
                      retrieved_card_ids:
                        type: array
                        items:
                          type: string
                      retrieval_count:
                        type: integer
                  runtime:
                    type: object
                    properties:
                      service:
                        type: string
                      version:
                        type: string
                      degraded:
                        type: boolean
        '401':
          description: Unauthorized request due to missing or invalid api key

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Enter your SOLOMON_ACTIONS_API_KEY value.